Working through index management issues
