+++ module/model-elasticsearch/index.ts
@@ -1,4 +1,3 @@
 export * from './src/config';
-export * from './src/schema';
-export * from './src/query-builder';
+export * from './src/util';
 export * from './src/source';
